chore: preparar servidor para deploy
