+++ package.json
@@ -2,10 +2,10 @@
   "name": "clube-vantagens-api",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node app.js",
-    "dev": "nodemon app.js",
+    "start": "node server.js",
+    "dev": "nodemon server.js",
     "test:api": "node ./tests/ping.js"
   },
   "keywords": [],
